requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-firestore
streamlit
watchdog
pytest
